requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
redis==5.0.1
orjson==3.9.10
python-socketio==5.10.0
python-multipart==0.0.6
httpx==0.25.2
timezonefinder==6.2.0
pytz==2023.3
psutil==5.9.6
